move test setup into removable directory
