+++ api/src/db/__test-setup__/initialize-test-db.js
@@ -1,4 +1,4 @@
-import { connect, loadModels, getDb } from './db-connector'
+import { connect, loadModels, getDb } from '../db-connector'
 
 afterAll(() => {
   const db = getDb()
